requirements.txt
Django
django-q2
croniter
requests
python-dotenv
jinja2
pytz
